+++ test/Service/TestAsset/AbstractMandatoryOptionsFactory.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Sake
+ *
+ * @link      http://github.com/sandrokeil/EasyConfig for the canonical source repository
+ * @copyright Copyright (c) 2014 Sandro Keil
+ * @license   http://github.com/sandrokeil/EasyConfig/blob/master/LICENSE.txt New BSD License
+ */
+
+namespace SakeTest\EasyConfig\Service\TestAsset;
+
+use Sake\EasyConfig\Service;
+
+abstract class AbstractMandatoryOptionsFactory extends Service\AbstractConfigurableFactory implements
+    Service\MandatoryOptionsInterface
+{
+}