+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # EasyConfig CHANGELOG
 
+## 1.3.0 (2014-09-24)
+
+- Added MandatoryOptionsInterface to check for mandatory options automatically
+- Added empty Module.php e.g. for modules.zendframework.com
+
 ## 1.2.0 (2014-04-29)
 
 - Updated source to PSR-4